major db overhaul, breaking changes in prices/order/discount/payments/refunds

+++ database/migrations/2019_09_17_165224_category_product.php
@@ -15,7 +15,7 @@ class CategoryProduct extends Migration
     {
         Schema::create('category_product', function (Blueprint $table) {
             $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('category_id');
+            $table->unsignedSmallInteger('category_id');
         });
     }
